feat: show obedience report

+++ lib/services/country.service.dart
@@ -5,7 +5,7 @@ import 'package:http/http.dart' as http;
 import '../utils/constants.dart';
 
 class CountryService {
-  static Future<double> getTemperature({String city = "kigali"}) async {
+  static Future<int> getTemperature({String city = "kigali"}) async {
     final url =
         "https://api.openweathermap.org/data/2.5/weather?q=${city.toLowerCase()}&appid=$apiKey&units=metric";
 
@@ -13,8 +13,8 @@ class CountryService {
 
     if (response.statusCode == 200) {
       double temp = jsonDecode(response.body)["main"]["temp"];
-      return temp;
+      return temp.toInt();
     }
-    return 0.0;
+    return 0;
   }
 }
